+++ packages/react/src/components/Button/Button.stories.tsx
@@ -13,12 +13,6 @@ type Story = StoryObj<typeof Button>;
 const meta: Meta<typeof Button> = {
   title: 'Kjernekomponenter/Button',
   component: Button,
-  parameters: {
-    status: {
-      type: 'beta',
-      url: 'http://www.url.com/status',
-    },
-  },
 };
 
 export default meta;
@@ -44,7 +38,6 @@ const icon = (
   </svg>
 );
 
-// Render functions need to accept "args" for code preview to work properly
 const VariantsTemplate = () => {
   return (
     <>