+++ packages/react/src/components/Chip/Toggle/Toggle.tsx
@@ -8,7 +8,7 @@ import classes from './Toggle.module.css';
 
 export type ToggleChipProps = {
   /**
-   * Condition if the check mark icon should be displayed
+   * Enables check mark icon
    */
   checkmark?: boolean;
 } & ChipBaseProps;
@@ -19,7 +19,7 @@ export const ToggleChip = forwardRef<HTMLButtonElement, ToggleChipProps>(
       children,
       size = 'small',
       selected = false,
-      checkmark = false,
+      checkmark = true,
       ...rest
     }: ToggleChipProps,
     ref,