+++ packages/react/src/components/LegacyAccordion/icons/index.ts
@@ -1,2 +1,0 @@
-export { Arrow } from './Arrow';
-export { CircleArrow } from './CircleArrow';