+++ packages/react/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.23.0](https://github.com/digdir/designsystem/compare/@digdir/design-system-react@0.22.0...@digdir/design-system-react@0.23.0) (2023-08-16)
+
+### Bug Fixes
+
+- **Pagination:** :wheelchair: Fix 'a11y' error for Pagination next/prev buttons ([#726](https://github.com/digdir/designsystem/issues/726)) ([e0e6fa4](https://github.com/digdir/designsystem/commit/e0e6fa424c9d1af54546eb344e58105271c8ea63))
+
+### Features
+
+- **Checkbox, CheckboxGroup, FieldSet, RadioButton, RadioGroup:** :wastebasket: Change to Legacy ([#716](https://github.com/digdir/designsystem/issues/716)) ([b22d836](https://github.com/digdir/designsystem/commit/b22d836c51ffcb024331eb07109547b288151888))
+- **Checkbox, Radio, Fieldset:** :sparkles: Release new components ([#720](https://github.com/digdir/designsystem/issues/720)) ([7facb4d](https://github.com/digdir/designsystem/commit/7facb4d1c77deffedb350fc7ed3e4bceecd4ff75))
+- **Link:** Create the component ([#711](https://github.com/digdir/designsystem/issues/711)) ([ac0adcc](https://github.com/digdir/designsystem/commit/ac0adcc2992499fd850e7974a54a7e4b19dfdd06))
+
+### BREAKING CHANGE
+
+- **enums:** 🔥 Remove `enums` ([#721](https://github.com/digdir/designsystem/issues/721))
+- **LegacyAccordion:** 🔥 Remove `LegacyAccordion` ([#707](https://github.com/digdir/designsystem/issues/707))
+
+# [0.22.0](https://github.com/digdir/designsystem/compare/@digdir/design-system-react@0.21.0...@digdir/design-system-react@0.22.0) (2023-08-09)
+
+### Bug Fixes
+
+- **Accordion:** Support dynamic height ([#689](https://github.com/digdir/designsystem/issues/689)) ([6624ba1](https://github.com/digdir/designsystem/commit/6624ba1eb75007cf07227f74a633f78d0b718ce3))
+- **Alert:** :bug: Left border should now be positioned correctly inside the box ([#702](https://github.com/digdir/designsystem/issues/702)) ([5fa2aeb](https://github.com/digdir/designsystem/commit/5fa2aeb4a17e77f5fbb98fd8ac039bb4bed0a473))
+- **Chip.Removable:** :bug: Use correct button type to avoid default submit ([#694](https://github.com/digdir/designsystem/issues/694)) ([357cfe8](https://github.com/digdir/designsystem/commit/357cfe827460031cc5e03f0e98f61947210bf40c))
+
+### Features
+
+- **Button:** :recycle: Change default `size` to medium (instead of small`) ([#698](https://github.com/digdir/designsystem/issues/698)) ([ab238df](https://github.com/digdir/designsystem/commit/ab238df8edb86f0432fe59b0d653986c6d3f09b5))
+- **Checkbox:** ✨ New `Checkbox` component (unreleased) ([#681](https://github.com/digdir/designsystem/issues/681)) ([6b49301](https://github.com/digdir/designsystem/commit/6b4930169802a860bad518173e18186407a36f02))
+- **Chip.Toggle:** ♻️ Enabled `checkmark` as default when `selected` ([#693](https://github.com/digdir/designsystem/issues/693)) ([049460c](https://github.com/digdir/designsystem/commit/049460c1b027f1846e5bf675b4e6768d6320458a))
+- **Label:** :sparkles: Added new prop `weight` for setting font weight ([#699](https://github.com/digdir/designsystem/issues/699)) ([1641752](https://github.com/digdir/designsystem/commit/16417526f584339544129ba990891f3dabf66267))
+
 # [0.21.0](https://github.com/digdir/designsystem/compare/@digdir/design-system-react@0.20.0...@digdir/design-system-react@0.21.0) (2023-08-02)
 
 ### Bug Fixes